Fixed the CreateFile function
Added some global variables. for hostname, user@ipaddress(amazon), and key. The key is the location of "rego.pem".  For now, it is just "rego.pem", meaning the key must be in the same directory as the new.go. To change the location of the key, just change variable key to a path location(user/home/example/rego.pem)